feat: 공지사항 detail id 없을때 처리 연결 - 백엔드 컨텐트 없을때 throw error 수정- 목록버튼 , 컨텐트보더 공통으로 수정
