+++ src/components/common/contentBorder/ContentBorder.styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ContentBorder = styled.div`
+  width: 100%;
+  height: 0.5px;
+  background: ${({ theme }) => theme.color.placeholder};
+`;
