+++ src/components/common/contentBorder/ContentBorder.tsx
@@ -1,0 +1,5 @@
+import * as S from './ContentBorder.styled';
+
+export default function ContentBorder() {
+  return <S.ContentBorder></S.ContentBorder>;
+}
