+++ src/components/customerService/noticeDetail/bottom/NoticeDetailBottom.tsx
@@ -1,5 +1,6 @@
-import { ROUTES } from '../../../../constants/routes';
 import type { OtherNotice } from '../../../../models/customerService';
+import ContentBorder from '../../../common/contentBorder/ContentBorder';
+import ListButton from './button/ListButton';
 import OtherNoticeButton from './button/OtherNoticeButton';
 import * as S from './NoticeDetailBottom.styled';
 
@@ -14,7 +15,7 @@ export default function NoticeDetailBottom({
 }: NoticeDetailBottomProps) {
   return (
     <S.Container>
-      <S.ContentBorder></S.ContentBorder>
+      <ContentBorder />
       {prev !== null ? (
         <OtherNoticeButton
           navigation='이전'
@@ -25,7 +26,7 @@ export default function NoticeDetailBottom({
       ) : (
         <S.NotOtherNotice>이전 공지사항이 없습니다.</S.NotOtherNotice>
       )}
-      <S.ContentBorder></S.ContentBorder>
+      <ContentBorder />
       {next !== null ? (
         <OtherNoticeButton
           navigation='다음'
@@ -36,12 +37,7 @@ export default function NoticeDetailBottom({
       ) : (
         <S.NotOtherNotice>다음 공지사항이 없습니다.</S.NotOtherNotice>
       )}
-      <S.ContentBorder></S.ContentBorder>
-      <S.ListWrapper>
-        <S.ListLink to={`${ROUTES.customerService}/${ROUTES.notice}`}>
-          <S.ListTitle>목록</S.ListTitle>
-        </S.ListLink>
-      </S.ListWrapper>
+      <ListButton />
     </S.Container>
   );
 }
