+++ src/components/customerService/noticeDetail/content/NoticeDetailContent.styled.ts
@@ -61,11 +61,3 @@ export const ContentWrapper = styled.div`
 `;
 
 export const Content = styled.p``;
-
-export const ContentBorder = styled.div`
-  width: 100%;
-  height: 0.5px;
-  background: ${({ theme }) => theme.color.placeholder};
-  position: relative;
-  z-index: 1;
-`;
