+++ src/pages/customerService/notice/Notice.styled.ts
@@ -21,11 +21,3 @@ export const Wrapper = styled.div`
 `;
 
 export const NoticeDetailLink = styled(Link)``;
-
-export const ContentBorder = styled.div`
-  width: 100%;
-  height: 0.5px;
-  background: ${({ theme }) => theme.color.placeholder};
-  position: relative;
-  z-index: 1;
-`;
